Refactor avl.zig to use 'const' instead of 'var' to be in line with zig versions 0.12 onwards (#10)
* Refactor avl.zig to use 'const' instead of 'var'

In the avl.zig file, occurrences of 'var' have been replaced with 'const' where possible and whitespace has been cleaned up for better readability. This change reduces the potential for accidental mutation of variables and enhances the maintainability of the codebase.

* Update source file paths and add library usage info in README

The build.zig file has been updated to modify the way source file paths are provided, now using the .src_path format to specify the owning build and sub path. Also, a minor update is made to the README file to inform users that they need at least Zig 0.13.x to use this library.

* fix(avl.zig): remove TABS in comments to compile under >=v0.14.7xx